A shape util
