+++ waveblocks/wavepackets/hawp_gradient_evaluator.hpp
@@ -1,6 +1,5 @@
 #pragma once
 
-#include <vector>
 #include <array>
 
 #include <Eigen/Core>
